move load_aztk_spark_config function to config.py
